Add getPersonFromDB to ID-service abstraction
We add a function 'getPersonFromDB' to the ID-service abstraction
'idManager' to abstract the '/getUser/' interface of the ID service.

This new method enables the user to query a person from the database by
passing the person's ID. This way, the local cache of the idManager
abstraction is bypassed and the full information on the person stored in
the DB is returned (instead of the cached and, thus, short version).

Perform small refactorings to perform the task and some code formatting.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>