Fix commit–artifact mapping for commit_dependency table
When inserting artifact information into the commit_dependency table,
the information is mapped to a specific commit. But, as a commit can
occur several times in a Git tree (via git-am, for example), it is
important to incorporate the current release range in the query for
finding the correct commit ID. Otherwise, the commit_dependency entries
may be mapped to the wrong commit ID.

To provide a solution for this, the method 'getCommitId' in the
DBManager is now able to handle (optionally given) release ranges.

Props to @bockthom for pointing this out.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>