Update cppstats to latest version (v0.9.4)
The latest version of cppstats fixes problems with dependencies and,
thus, also with the Codeface installation.

Previously, cppstats depended on the Python package 'enum'.
Unfortunately, this package clashes with the package 'enum34' which
itself is a dependency of the package 'jira' introduced in commit
5ba4c1bc8b60fe3dd7e50860a26393ef1c930f1c. With the new version, cppstats
was updated to use 'enum34' as well.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>