Fix R package installation for package 'markovchain'
As the newest version of the R package 'markovchain' is not compatible
with R version < 3.5, we need to manually need to make sure that a
compatible version for lower versions (as on TravisCI) is installed. See
also https://travis-ci.org/se-passau/codeface/builds/421118750.

Additionally, the flag 'quiet' on R package installations is set to
'FALSE' to ensure that installation problems can be identified earlier.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>