Replace email address <noreply@github.com>
As the email address <noreply@github.com> is used with multiple author names,
this email address potentially leads to combining multiple authors to one
person in the database. To omit that, don't use this email address, but
replace it with an email address which has the author's name as additional
prefix. For example, 'Hans Huber <noreply@github.com>' is replaced by
'Hans Huber <Hans.Huber.noreply@github.com>'.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>