Fix encoding problems for persons' names
As the codeface database only can handle utf-8 encoded strings using 3 bytes,
we have problems if there is a character which needs 4 bytes in utf-8
encoding. As we alredy have a function 'encode_as_utf8' for text encodings
etc., which replaces such characters by a space, we can reuse this function
also for persons' names right before inserting them or querying them in the
database.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>