[conway] Add rate limiting for JIRA requests in function 'parse_jira_issues'
As the Infrastructure Team of the Apache Server Foundation allows only
50,000 JIRA requests per 24 hours, we need to restrict the number of
JIRA requests when parsing JIRA issues. Therefore, we add a counter for
JIRA requests and if this counter exceeds 45,000 calls, we preventively
pause for 24 hours before resetting the counter and continuing with the
next JIRA request.

Note: The counter only considers the requests within one run of
'parse_jira_issues'. So, make sure that you don't run several
instances of 'parse_jira_issues' in parallel or in sequence.
In addition, we don't count the requests to the JIRA API that are
sent via titan to get the weekly xml files containing the issue data.
(Approximately, titan may send one request per week for which the JIRA
issue data should be retrieved.) Also other tools may send JIRA
requests, which are also not considered here. Hence, just taking care
of the requests of 'jira_issue_processing.py' is not enough, you may
need to manually decrease the maximum number of requests if other
tools might have sent JIRA requests within the last 24 hours.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>