Update ctags version (use universal-ctags version Sep 3 2021, 18:12:18)
To extract functions correctly, for some programming languages, we need to
configure ctags accordingly. However, exuberant-ctags from 2009 is not up to
date any more. Specifically, we need to update to universal-ctags from 2021
(e.g., to be able to extract functions in R correctly).

However, universal-ctags is not availabe on Ubuntu versions prior to 20.04.
Therefore, we need to update the operating system to the most recent Ubuntu
version, in our case Ubuntu 22.04.

Upgrading to Ubuntu 22.04., in addition, involves lots of other dependency
problems, which are also fixed in this commit.

In particular, we needed to:
- replace exuberant-ctags by universal-ctags
- remove system libraries which are not availabe any more from
'install_common.sh'
- replace python package 'mysql-python' by 'mysqlclient'
- replace python package 'python-ctags' by 'python-ctags3'

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>