Update R packages to be compatible with R version 4.1
In line with updating the operating system to Ubuntu 22.04, we also updated the
R version from 3.x to 4.1. As a consequence, some R package installations and
versions need to be adjusted accordingly.

markovchain: Version 0.6.9.11 is not applicable to R 4.1. Use the standard
version instead.

Some of the pinned R packages are not compatible with newer versions of package
BH. Therefore, we need to pin BH to version 1.75.0-0. Consequently, package
rjson needs to be pinned to version 0.2.20.

logging: The logging statements used by the R part of Codeface rely on logging
package 0.8-104. More recent versions of the logging package contain braking
changes, resulting in empty log statements or log statements not being printed
at all. Therefore, we need to pin package logging to version 0.8-104.

Bioconductor packages: The way on how to install packages from bioconductor has
changed substantially. Use BiocManager instead of biocLite.

Package "tm": Several updated packages are not compatible with the pinned
version of package tm any more. Therefore, we need to update "tm" to its most
recent version, without any consequences on the fixed package tm.plugin.mail.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>