Adjust ctags configuration for several programming lanugages
As a follow-up for commit 8124c96657dfa2eae896bfcef7bb72195e5b816a, and since we
are now able to use a more recent version of ctags, the ctags kind
configurations of interest are adjusted for some of the programming languages
(Go, Ruby, Dart, Perl, R).

The default is still to use doxygen. Only when we have already experienced that
doxygen is not able to extract functions for a specific programming language, we
switch to ctags. In this commit, we do this for the languages Rust and R.

ATTENTION: If a programming language is analyzed that was not analyzed in any
project before, it may be necessary to check whether doxygen correctly extracts
functions for these languages and if not, use ctags instead (i.e., remove it
from the list of file endings used for doxygen parsing). It may also be
necessary to add additional kinds specific for this programming language to the
ctags configuration. Details about the kinds for specific programming languages
can be found in the implementation of the ctags parsers:
https://github.com/universal-ctags/ctags/tree/master/parsers

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>