Get rid of error message that commit diffstats could not be parsed
As of more recent git versions, "git show --format=full --shortstat --numstat"
has another output format if a commit is empty (or contains just whitespace
changes in case of "--ignore-space-change"). In particular, the last line,
stating how much files were changed, how many inserations and how many
deletions, is missing in the new output of git. As a consequence, the
functionality to detect empty commits in Codeface does not work any more and
results in printing the error "Could not parse diffstat". However, as this is
not an error, we get rid of this error by adjusting the functionality to detect
empty commits, to cover the current and past output format of git:

Instead of checking whether the last line starts with 4 spaces, we also check
whether the last line is empty, and in such a case, check whether the second
last line starts with 4 spaces. Using that, the above described error is not
printed any more.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>