Handle doxygen bug
In some cases, doxygen returns "-1" as the start line of a function or the end
line of a function. However, this is an invalid value due to some wrong
identification in doxygen. If we use "-1" as start line of a function, this
leads to problems in Codeface, as "-1" is the default value for "File_Level",
ending up in an error caused by inconsistent usage of "-1" then. To get rid of
this error, check if the doxygen output is "-1", and if so, just ignore this
function and print a warning instead.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>