Fix loss of database connection after timeout
The new environment which uses the mysqlclient package has some different
behavior than before: If a database connection is closed by the database server
due to server connection timeout, calling "ping" does not reestablish the
connection again (as it did before). Instead, we need to close the connection
completely and open a completely new connection using exactly the same
parameters. Therefore, store these parameters in a fields that can be accessed
during opening a new connection (as previously only the constructor knew about
them) and provide a function for opening a connection using these parameters.

´Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>