Do not analyze functions in SQL files (workaround, not a fix)
The DBAnalysis module is present to determine functions in SQL files, but the
DBAnalysis module uses unavailable imports: pytest  is not installed and
therefore cannot be imported (but pytest may not be needed anyway), and, more
importantly, it is completely unknown where 'parse_tablename',
'parse_databasename', and 'parse_indexname' come from. As these imports cannot
be resolved, the complete module DBAnalysis cannot be importet and therefore not
used. To prevent ending up in an error when trying to analyze functions in SQL
files, print a warning and return an empty list of function lines.

NOTICE: This commit is not a fix, it just prevents raising an exception instead
of fixing the actual import problem!

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>