Update cluster scripts: Adjust calls to BoDeGHa for ssh-based URLs
Passing the correct parameter containing organization name and repo name to
BoDGHa did not work for repos cloned via ssh as the URL string processing was
specific to https cloned repos. This is fixed by adding an additional check:
If the organization name contains a @ and a : (as in git@github.com:reponame),
then everything up to the first colon is removed from the reponame.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>