Fix handling of missing commit message
If a commit has no commit message, the corresponding parsing function for the
commit message fails with an index-out-of-bounds error, as the respective part
which is accessed does not exist in such a case.

To fix this, check for the existence of a commit's commit messages and print a
warning if there is no commit message. In such a case, set the number of lines
as well as the number of characters of the commit message to zero.

In addition, this commit fixes another broken warning message.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>