Handle commit authors of special form
Some projects use a special format of (name, e-mail) pairs in their commits:
    address@domain.tld <>
Up until now, codeface was not able to extract the e-mail address from such a
format. With this commit, the above stated format is replaced by the
following format:
    address AT domain.tld <address@domain.tld>
This replacement makes it possible to parse the e-mail address correctly in such
a special case.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>