Add workaround for bug in 'git log --before' + extract earliest root commit
Due to a bug in git, broken author information in commit objects can lead to a
timestamp of 0 when using the --before option although the dates themselves are
not broken and can be parsed without problems. For more details, see the whole
thread conversation on the git mailing list here:
https://lore.kernel.org/git/7728e059-d58d-cce7-c011-fbc16eb22fb9@cs.uni-saarland.de/

To avoid running into an infinite while loop when creating Codeface's threemonth
ranges (due to timestamps being 0), check if the start date of a range is
earlier than the date of the earliest root commit and break if this is the case.
In this case, use the date of the earliest root commit as start date instead.

This also fixes another bug: In the previous implementation, it was assumed that
the initial commit in the repository also was the first commit (i.e., the commit
with earliest time stamp). However, this does not need to be the case, as other
branches can be merged into the main branch later, which can have a root commit
that has an earlier time stamp than the initial commit in the repository.
Therefore, use the earliest root commit for the initial range instead of the
intial commit in the version history.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>